# server/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AppStoreCat API",
        "description": "Open source app intelligence toolkit API. Provides endpoints for app registration, DNA building, store listing tracking, and analytics.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:7460/api/v1",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/apps": {
            "get": {
                "tags": [
                    "Apps"
                ],
                "summary": "List tracked apps",
                "operationId": "listApps",
                "parameters": [
                    {
                        "name": "platform",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "ios",
                                "android"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of tracked apps",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/AppResource"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Apps"
                ],
                "summary": "Register and track an app",
                "operationId": "storeApp",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreAppRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "App registered",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AppDetailResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/apps/{platform}/{externalId}": {
            "get": {
                "tags": [
                    "Apps"
                ],
                "summary": "Get app details",
                "operationId": "showApp",
                "parameters": [
                    {
                        "name": "platform",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "ios",
                                "android"
                            ]
                        }
                    },
                    {
                        "name": "externalId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "App details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AppDetailResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/apps/{platform}/{externalId}/listing": {
            "get": {
                "tags": [
                    "Apps"
                ],
                "summary": "Get store listing for a specific country",
                "operationId": "appListing",
                "parameters": [
                    {
                        "name": "platform",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "ios",
                                "android"
                            ]
                        }
                    },
                    {
                        "name": "externalId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "country",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "tr"
                        }
                    },
                    {
                        "name": "language",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "tr"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Store listing",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ListingResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/apps/{platform}/{externalId}/track": {
            "post": {
                "tags": [
                    "Apps"
                ],
                "summary": "Track an app",
                "operationId": "trackApp",
                "parameters": [
                    {
                        "name": "platform",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "ios",
                                "android"
                            ]
                        }
                    },
                    {
                        "name": "externalId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "App tracked"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Apps"
                ],
                "summary": "Untrack an app",
                "operationId": "untrackApp",
                "parameters": [
                    {
                        "name": "platform",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "ios",
                                "android"
                            ]
                        }
                    },
                    {
                        "name": "externalId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "App untracked"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "CompetitorRelationship": {
                "type": "string",
                "example": "direct",
                "enum": [
                    "direct",
                    "indirect",
                    "aspiration"
                ]
            },
            "Platform": {
                "type": "string",
                "example": "ios",
                "enum": [
                    "Ios",
                    "Android"
                ]
            },
...
            "StoreAppRequest": {
                "required": [
                    "external_id",
                    "platform"
                ],
                "properties": {
                    "external_id": {
                        "type": "string",
                        "example": "389801252"
                    },
                    "platform": {
                        "type": "string",
                        "example": "ios",
                        "enum": [
                            "ios",
                            "android"
                        ]
                    }
                },
                "type": "object"
            },
...
            "AppDetailResource": {
                "required": [
                    "id",
                    "name",
                    "platform",
                    "external_id",
                    "created_at"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Instagram"
                    },
                    "platform": {
                        "$ref": "#/components/schemas/Platform"
                    },
                    "external_id": {
                        "type": "string",
                        "example": "389801252"
                    },
                    "publisher": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/Publisher"
                            }
                        ],
                        "nullable": true
                    },
                    "category": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/StoreCategory"
                            }
                        ],
                        "nullable": true
                    },
                    "supported_locales": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "EN",
                            "DE",
                            "FR"
                        ],
                        "nullable": true
                    },
                    "original_release_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2010-10-06",
                        "nullable": true
                    },
                    "is_free": {
                        "type": "boolean",
                        "example": true
                    },
                    "icon_url": {
                        "type": "string",
                        "nullable": true
                    },
                    "rating": {
                        "type": "number",
                        "format": "float",
                        "example": 4.68,
                        "nullable": true
                    },
                    "rating_count": {
                        "type": "integer",
                        "example": 31,
                        "nullable": true
                    },
                    "version": {
                        "type": "string",
                        "example": "1.2.4",
                        "nullable": true
                    },
                    "file_size_bytes": {
                        "type": "integer",
                        "example": 73822208,
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "listings": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ListingResource"
                        }
                    },
                    "versions": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/VersionResource"
                        }
                    },
                    "changes": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/StoreListingChange"
                        }
                    },
                    "competitors": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/CompetitorResource"
                        }
                    }
                },
                "type": "object"
            },
...
            "AppResource": {
                "required": [
                    "id",
                    "name",
                    "platform",
                    "external_id",
                    "created_at"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Instagram"
                    },
                    "platform": {
                        "$ref": "#/components/schemas/Platform"
                    },
                    "external_id": {
                        "type": "string",
                        "example": "389801252"
                    },
                    "publisher": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/Publisher"
                            }
                        ],
                        "nullable": true
                    },
                    "category": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/StoreCategory"
                            }
                        ],
                        "nullable": true
                    },
                    "icon_url": {
                        "type": "string",
                        "nullable": true
                    },
                    "rating": {
                        "type": "number",
                        "format": "float",
                        "example": 4.68,
                        "nullable": true
                    },
                    "rating_count": {
                        "type": "integer",
                        "example": 31,
                        "nullable": true
                    },
                    "version": {
                        "type": "string",
                        "example": "1.2.4",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "CompetitorResource": {
                "required": [
                    "id",
                    "relationship",
                    "app"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "relationship": {
                        "$ref": "#/components/schemas/CompetitorRelationship"
                    },
                    "notes": {
                        "type": "string",
                        "nullable": true
                    },
                    "app": {
                        "$ref": "#/components/schemas/AppResource"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "ListingResource": {
                "allOf": [
                    {
                        "$ref": "#/components/schemas/StoreListing"
                    }
                ]
            },
...
            "VersionResource": {
                "allOf": [
                    {
                        "$ref": "#/components/schemas/AppVersion"
                    }
                ]
            },
...
            "AppVersion": {
                "required": [
                    "id",
                    "version"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "version": {
                        "type": "string",
                        "example": "1.2.4"
                    },
                    "release_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2026-01-15",
                        "nullable": true
                    },
                    "whats_new": {
                        "type": "string",
                        "nullable": true
                    },
                    "file_size_bytes": {
                        "type": "integer",
                        "example": 73822208,
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "Publisher": {
                "required": [
                    "id",
                    "name",
                    "platform"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Instagram, Inc."
                    },
                    "external_id": {
                        "type": "string",
                        "example": "389801255",
                        "nullable": true
                    },
                    "platform": {
                        "$ref": "#/components/schemas/Platform"
                    },
                    "url": {
                        "type": "string",
                        "example": "https://apps.apple.com/us/developer/instagram-inc/id389801255",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "StoreCategory": {
                "required": [
                    "id",
                    "name",
                    "slug",
                    "platform"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Photo & Video"
                    },
                    "slug": {
                        "type": "string",
                        "example": "photo-video"
                    },
                    "platform": {
                        "$ref": "#/components/schemas/Platform"
                    }
                },
                "type": "object"
            },
            "StoreListing": {
                "required": [
                    "id",
                    "language"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "version_id": {
                        "type": "integer",
                        "nullable": true
                    },
                    "language": {
                        "type": "string",
                        "example": "tr"
                    },
                    "title": {
                        "type": "string",
                        "example": "Instagram"
                    },
                    "subtitle": {
                        "type": "string",
                        "example": "Photos & Videos",
                        "nullable": true
                    },
                    "description": {
                        "type": "string"
                    },
                    "whats_new": {
                        "type": "string",
                        "nullable": true
                    },
                    "icon_url": {
                        "type": "string",
                        "nullable": true
                    },
                    "screenshots": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "nullable": true
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 6.99
                    },
                    "currency": {
                        "type": "string",
                        "example": "USD",
                        "nullable": true
                    },
                    "video_url": {
                        "type": "string",
                        "nullable": true
                    },
                    "fetched_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "StoreListingChange": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "version_id": {
                        "type": "integer",
                        "nullable": true
                    },
                    "language": {
                        "type": "string",
                        "example": "us"
                    },
                    "field_changed": {
                        "type": "string",
                        "example": "description"
                    },
                    "old_value": {
                        "type": "string",
                        "nullable": true
                    },
                    "new_value": {
                        "type": "string",
                        "nullable": true
                    },
                    "detected_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Authentication using Bearer token. Example: Bearer {your-token}",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}